requirements.txt
Database==0.6.0
discord.py==1.5.1
discord==1.0.1
imgurpython==1.1.7
python-dotenv==0.15.0
